chore: rename src/util/vue to src/util/vue-compat
@vue/test-utils@2 seems to have an issue with importing from that path. All the imported are undefined. When I change the path, it works fine. It doesn't happen with VTU@1, so I assume it's a bug from VTU@2, but let's resolve it this way for now.
