+++ jest.setup.js
@@ -1,5 +1,5 @@
 import createSerializer from 'jest-serializer-html/createSerializer';
-import { isVue2, Vue2 } from './src/util/vue';
+import { isVue2, Vue2 } from './src/util/vue-compat';
 
 if (isVue2) {
   Vue2.config.productionTip = false;
