+++ src/util/createInstantSearchComponent.js
@@ -1,7 +1,7 @@
 import { createSuitMixin } from '../mixins/suit';
 import { version } from '../../package.json'; // rollup does pick only what needed from json
 import { _objectSpread } from './polyfills';
-import { isVue3, version as vueVersion } from '../util/vue';
+import { isVue3, version as vueVersion } from '../util/vue-compat';
 
 export const createInstantSearchComponent = component =>
   _objectSpread(
