+++ src/util/vue-compat/index.js
@@ -1,0 +1,1 @@
+export * from './index-3';
